backend/requirements.txt
openai-agents
fastapi
uvicorn
pydantic
supabase
redis
umap-learn 